+++ frontend/src/.eslintrc
@@ -5,6 +5,7 @@
     "prettier/prettier": ["error"],
     "react/prop-types": 0,
     "jsx-a11y/no-static-element-interactions": 0,
-    "jsx-a11y/click-events-have-key-events": 0
+    "jsx-a11y/click-events-have-key-events": 0,
+    "endOfLine": "auto"
   }
 }
