--- frontend/src/components/Header.js
@@ -1,27 +1,0 @@
-import styles from './Header.module.css'
-import { NavLink } from 'react-router-dom'
-import { FaUser as Account } from 'react-icons/fa'
-import { ImEnter as SignIn } from 'react-icons/im'
-import AuthService from '../utils/AuthService'
-
-function Header() {
-    const loggedIn = !!AuthService.getCurrentUser()
-
-    return (
-        <div className={styles.header}>
-            <NavLink to={'/home/0'} className={styles.headerLinks} activeStyle={{ color: '#0c1e66' }}>
-                Home
-            </NavLink>
-            <span className={styles.title}>askIBU</span>
-            <NavLink
-                to={loggedIn ? '/account' : '/login'}
-                className={styles.icons}
-                activeStyle={{ color: '#0c1e66' }}
-            >
-                {loggedIn ? <Account /> : <SignIn />}
-            </NavLink>
-        </div>
-    )
-}
-
-export default Header
