+++ frontend/src/components/Header/Header.jsx
@@ -1,9 +1,8 @@
-import React from "react";
+import styles from "./Header.module.css";
 import { NavLink } from "react-router-dom";
 import { FaUser as Account } from "react-icons/fa";
 import { ImEnter as SignIn } from "react-icons/im";
 import AuthService from "../../utils/AuthService";
-import styles from "./Header.module.css";
 
 function Header() {
   const loggedIn = !!AuthService.getCurrentUser();
@@ -11,7 +10,7 @@ function Header() {
   return (
     <div className={styles.header}>
       <NavLink
-        to="/home"
+        to={"/home/0"}
         className={styles.headerLinks}
         activeStyle={{ color: "#0c1e66" }}
       >
